gadget.yaml: add system-recovery partition
For UC20 we want to have a recovery partition. This adds to
the existing system-boot partition. The size is currently set
to 600Mb - this is a bit arbitrary and will allow us roughtly
2 recovery systems. We have the following sizes today:
 * core18:  54MB
 * snapd:   23MB
 * gadget:   1MB
 * kernel: 200MB
The recovery partition is setup after the system-boot
partition.

It also adds a initial grub.cfg which is very basic and
only supports a single recovery system at this point but
that recovery system can be selected via the grubenv:
- `snap_recovery_kernel`
- `snap_recovery_core`
settings.