grub.cfg-normal: go back to the UC16 way of loading the kernel
In run mode this will go back to the old way of loading the kernel
but from the "ubuntu-data" partition. This is just a step on the
way so that we can create integration tests for image creation
and seeding. Soon this will be replaced with the UC20 single-file
(no grubenv) way of booting in "run" mode.