grub.cfg-boot: use $prefix/kernel instead of $ubuntu-boot/kernel
Using $prefix/kernel allows us at runtime to not look at the root directory of
the ubuntu-boot partition and just use grub as-is mounted at /boot/grub.

Also change the name of the menuentry when we do boot an old kernel through the
temporary compatibility mode. This will be removed entirely when snapd is
updated to extract kernel assets properly.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>