grub: boot configuration is shipped and managed by snapd
Since v2.46, snapd can ship and manage the boot configuration for grub for both
the run mode and recovery. Drop the recovery mode config and truncate grub.conf
so that we are left with a suitable marker for snapd. Also, remove references to
grub*.conf from gadget.yaml, as there is no need to list those files in content
section anymore.

Signed-off-by: Maciek Borzecki <maciek.borzecki@gmail.com>