snapcraft.yaml: allow building outside of launchpad with old binaries
The measurements in 2.06 grub are currently incorrect, force using
focal binaries, without requirement to specify archive in launchpad
/+snap/ page such that one can build the gadget locally correctly too.

Also remove all unused dependencies. And it is no longer required to
stage build-deps, instead of build-depend on them, as they are no
longer conflicting with anything installed on the host.

Fixes #57