snap: increase size of ubuntu-save to 64Mb
The ubuntu-save partition needs to be bigger than 32Mb for devices
that have 4k sector sizes. The gadget does not know in advance
on what physical sector size it will run so increasing the size
for everything makes most sense.

The reason this is needed is that for devices <= 32Mb snapd will
format them with a logical sector size of 1Kb to avoid wasting
too much space for the journal (see osutil/mkfs/mkfs.go:mkfsExt4).
However on encrypted volumes this sectorsize can be incompatible
with what dm-crypt selected for the logical sector size and
when that does not match mkfs fails as observed in LP:#1968356.

This size increase fixes the install failure on devices with
physical sector size of 4k.