Build from snapcraft.yaml and use ucdev PPAs
Remove makefile and do special steps from an override-build section
instead.

Take now signed grub and shim form the ucdev/uc-staging-ppa PPA, using
package-repositories. Now we use epochs in the version of
grub-efi-amd64-signed and shim-signed packages, so the build should
always take the ones from the PPA instead of those in the archive, as
2:<version> will always be greater than the package version in the
latter.