Use chdist to use clean sources and honor SERIES
"apt-get download" would use whatever is in sources.list; instead, setup
an APT environment for target architecture and series. This means the
APT config is now controlled and actually honors target series.

pull-lp-debs would be much more elegant in achieving this, but it's not
in main, so is likely to be problematic when called from livecd-rootfs
or in clean main-only environments.